IEEE Conversions to Vue
